+++ lizard_worker/management/commands/start_scenario_new.py
@@ -4,13 +4,13 @@
 from optparse import make_option
 
 from django.core.management.base import BaseCommand
-from flooding_worker.file_logging import setFileHandler, removeFileHandlers
-from flooding_worker.worker.action_workflow import ActionWorkflow
-from flooding_worker.worker.broker_connection import BrokerConnection
-from flooding_worker.worker.message_logging_handler import AMQPMessageHandler
+from lizard_worker.file_logging import setFileHandler, removeFileHandlers
+from lizard_worker.worker.action_workflow import ActionWorkflow
+from lizard_worker.worker.broker_connection import BrokerConnection
+from lizard_worker.worker.message_logging_handler import AMQPMessageHandler
 
 import logging
-log = logging.getLogger("flooding.management.start_scenario")
+log = logging.getLogger("worker.management.start_scenario")
 
 
 class Command(BaseCommand):