+++ CHANGES.rst
@@ -6,3 +6,7 @@ Changelog of lizard-worker
 ----------------
 
 - Initial project structure created with nensskel 1.26.
+
+- Transform flooding-worker to lizard-worker.
+
+- Move tasks and perform_task.py to flooding-lib.